+++ script/api/api_settings.cc
@@ -44,7 +44,7 @@ bool get_time_based_routing_enabled(settings_t* settings, const goods_desc_t* go
 bool set_time_based_routing_enabled(settings_t* settings, const goods_desc_t* goods, bool enabled)
 {
 	if(  settings  &&  goods  ) {
-		const uint8 catg_index = goods->get_catg_index();
+		const uint16 catg_index = goods->get_catg_index();
 		if(  settings->get_time_based_routing_enabled(catg_index) != enabled  ) {
 			settings->set_time_based_routing_enabled(catg_index, enabled);
 			welt->set_schedule_counter();
